// packages/backend/src/users/dto/create-user.dto.d.ts
export declare class CreateUserDto {
    email?: string;
    phone?: string;
    username: string;
    password: string;
    nickname?: string;
    avatar?: string;
    roleId?: string;
    phoneVerified?: boolean;
    emailVerified?: boolean;
    wechatId?: string;
    provider?: string;
}
//# sourceMappingURL=create-user.dto.d.ts.map