// packages/backend/src/users/dto/update-user.dto.d.ts
import { UserStatus } from '@prisma/client';
export declare class UpdateUserDto {
    email?: string;
    phone?: string;
    username?: string;
    password?: string;
    nickname?: string;
    avatar?: string;
    roleId?: string;
    status?: UserStatus;
}
//# sourceMappingURL=update-user.dto.d.ts.map